Fix mouse coords for programs interested with it
Summary:
In normal usage Konsole offsets the position of the mouse
by half a char so that the user can extend the selection to
the last column. When using programs interested with
mouse (vim, mc, aptitude, etc...) Konsole shouldn't apply
the offset since mouse events should always apply to the
character cell under the pointer. This patch sets the 'edge'
flag of getCharacterPosition to !_usesMouseTracking.

Reviewers: hindenburg, ahmadsamir

Reviewed By: hindenburg

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D13089