Return early - minimize if indentation
Summary:
Use initializer list to create QSet

Don't check what doesn't needs to be checked.

the QBitArray knows how to deal with true / false
directly, so remove the if's. the speedgain of the
if's are probabely slower than just setting the values
to true / false again anyway as we are not suffering
from a possible branch-prediction, and plus, code is
cleaner.

No need to test for true / false to set true false

Just set it directly

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D13512