Remove SearchBar from the MainWindowOverview
Summary:
Move the SearchBar to TerminalDisplay, killing the dependency from
the MainWindow and allowing it to be positioned as an overlay.
This fixes the following bugs:

- Opening and closing the SearchBar while top/htop is open
would garble the screen as top would wrongly redraw
two rows below, filling the screen with wrong text
- with tabs and splits, opening two searchbars and
clicking in the close button of the first without
focusing the terminal display first would cause
the tabbar not to close
- Smaller and prettier.

This never belonged here.
Searchbar belongs to the TerminalDisplay
Correctly position the SearchBar on topLeft
Fix initial size of the SearchBar
Icons, Position and Borders
Change the icons, positions and borders
of the SearchBar
Fix fill background of the Search Tab
Don't show the menubar by default
Fix icons & text

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: ahmadsamir, ngraham, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D13511