Add reset font size shortcut
Summary:
Whenever I change the font size doing so is usually temporary and I usually go back to the default font size in after a short time. Currently I do so by changing to a different profile and discarding the change, which resets the font size to the default. I feel like there should be a shortcut for resetting the size.

I could not find an appropriate icon, so I left it without icon.
The default shortcut ctrl+r seemed reasonable and easy to type, 'r' reset.

Test Plan: Manual.

Reviewers: hindenburg, tcanabrava, #vdg, ngraham

Reviewed By: #vdg, ngraham

Subscribers: muhlenpfordt, ach, kvermette, rizzitello, lbergdoll, hindenburg, fabianr, dhaumann, kde-doc-english, z3ntu, ngraham, konsole-devel

Tags: #konsole, #documentation

Differential Revision: https://phabricator.kde.org/D15380