Check clipboard->mimeData() is valid
Summary:
BUG: 395666

Bug reported crashes on yakuake in a wayland session. Further
investigation revealed that other KDE product Kolourpaint experienced
nullptr on wayland while getting MimeData which was consistent with
the stacktrace provided by the author of the bug.

This diff validates the presence of MimeData before proceeding with
the drag action.

Test Plan:
1. Use wayland session
2. Open Yakuake in fullscreen
3. With a touch enabled laptop, scroll with many fingers (possibly 8)
4. Yakuake should not crash

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: anthonyfieroni, ngraham, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D15794