Don't consider a colon before whitespace to be part of a word
Summary:
Currently, colons are considered to be part of words for the purpose of
double-click selection. This allows, for example, URLs to be selected
with a double click. This is good.

However, including colons works less well for paths that end with a colon
and whitespace--such as `grep` output. Removing the colon from the list
of word characters isn't a great solutionsince then URLs can no longer be
selected by double-clicking.

This patch resolves that particular issue by always omitting a colon from
double-click selection when it's immediately followed by whitespace.

BUG: 399291
FIXED-IN: 18.12.0

Test Plan:
`grep -ri charClass .`
Then double-click on one of the paths in the output. The trailing colon
is not included in the selection.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D16578