[Terminal display] Limit the search result line highlight width
Summary:
The search result highlight was using the whole container width,
this creates some glitches when scrolling because the right-most
part of the line is not repainted.
Limiting the search highlight to the number of columns fixes this
and is more consistent with line selection highlight too.

Test Plan:
Highlight a line by searching something in the konsole, then scroll.
There should be no glitches left on the right.

{F6396359}

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, ngraham, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D16730