Properly close search when hitting Escape
Summary:
When performing a search and the focus is no longer on the
searchbox, hitting escape would hide the searchbox but would
not really end the search.

Test Plan:
Search something in the terminal, then click within konsole
but outside the searchbox so that it loses focus and hit
escape. The search highlightings should now be removed.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, ngraham, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D16754