Remove option to hide application name from titlebar
Summary:
Since the application name is added by Qt, it is no longer
possible to remove the trailing '- Konsole' so the GUI
option is obsolete.

BUG: 364391
FIXED-IN: 19.03

Test Plan:
Change settings for existing Konsole and notice that the checkbox
doesn't change anything. After applying this the checkbox should
be removed.

Reviewers: #konsole, konsole-devel, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17049