Allow small colour changes when varying dark backgrounds
Summary:
Black backgrounds don't get a (visible) random colour if "Vary the
background colour" has been checked. This is because the only
changes currently made aren't visible if the colour is black. Instead,
if we get a black background allow a small change in the colour
value so that the randomized colour is noticeable.

BUG: 311263
FIXED-IN: 18.12

Test Plan:
Create a new profile with a black background and the "Vary the
background colour" option checked. Opening new tabs with this
profile should produce tabs with slightly different background colours.

Reviewers: #konsole, konsole-devel, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: #konsole, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17045