Don't hide setFont, it's not virtual
Summary:
setFont is not a virtual method, so overriding it makes no sense.
if you have a pointer to the widget and call setFont things will
still (not) work, and if you have a pointer to the TerminalDisplay
and call setFont, nothing will happen, so just remove it.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17221