fix drawing box chars, avoid storing and saving state all the time
Summary:
to get the box chars to be drawn correctly we need to turn on high
quality antialiasing in qpainter. in addition only turn it on if
antialiasing is enabled.

lastly qpainter.save()/restore() is called very often, so try to avoid
that if it isn't necessary.

BUG: 401463

Test Plan:
`cat tests/boxes.txt`

old:

{F6428268}

new:

{F6450304}

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: wbauer, konsole-devel, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D16947