Don't duplicate the tabs when Spliting the View
Summary:
Instead of iterating over the sessions to fetch the view and
duplicate the current screen once for each split, ignore all
of them and just create a new terminal view. The only thing
this terminal view shares with the previous current widget
is the profile.

use createView instead of manually trying to create the view

Fixes bug while closing the split

Don't duplicate tabs in the splits

BUG: 385697
BUG: 380455
FIXED-IN: 19.04

Reviewers: thsurrel, hindenburg, gennad, ngraham, #konsole

Reviewed By: thsurrel, hindenburg, ngraham, #konsole

Subscribers: ngraham, hindenburg, gennad, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17267