fix cursor when anti aliasing is enabled
Summary:
When there is a margin of 1 pixel, but some of the anti-aliasing
"blur" leaks to 0,0. so it is fixed with a full update() (e. g. when
it regains focus), but not when just some content updates in the
window.   Without this it draws outside the content rect, most
visible as a vertical line if the cursor is completely to the left and
then moves.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: pbraun, konsole-devel, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17414

(cherry picked from commit e7085310d6d594823d0ed491fa8bdbd99dec4932)