Only show tab icon if the user has specifically chosen one
Summary:
This patch introduces a UI change: tabs only show an icon if it's not
the default one; that is to say, only if the user has actually customized
the icon. There are several reasons for this:
1. **Conceptual:** When only non-default profile icons show up in the tab,
  it does a better job of highlighting the fact that a non-default profile
  is being used.
2. **Aesthetic:** you no longer have the same icon in every tab by default,
  which looks bad and causes the eye to ignore it over time.
3. **Icon-specific:** the `utilities-terminal` icon doesn't look good
  against a dark background (https://bugs.kde.org/show_bug.cgi?id=367696)
4. **Code hygiene:** reduces a bit of redundancy in the session controller
   code.

BUG: 401864
CCBUG: 367696
FIXED-IN: 19.04.0

Test Plan:
Open multiple tabs and make one of them use a profile with a
non-default icon:
{F6502534, size=full}

Verify that bell signals still cause tabs to get icons.

Reviewers: #konsole, #vdg, ndavis, hindenburg

Reviewed By: #konsole, #vdg, ndavis, hindenburg

Subscribers: fabianr, abetts, hindenburg, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D17747