Fix undefined behavior and set QT_MIN_VERSION to 5.9.0
Summary:
Arguments in some QCOMPARE calls in KeyboardTranslatorTest had slightly
different types, what according to Qt documentation is not allowed.
The types of both QCOMPARE arguments must be exactly the same, otherwise
the code doesn't compile with some Qt versions.

With this fix it is safe to lower minimum Qt version to 5.9.0. Patch
number is incremented only for bugfix releases, so as long as invalid
code is not used everything should work.

The fix allows to compile Konsole using system libs on larger number
of systems, e.g. Ubuntu 18.04 LTS (Qt 5.9.5).

Tested with Qt 5.9.[0-5] and 5.11.1

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D19859