Add DesktopEntry to notifyrc
Summary:
This allows the notification KCM to identify Konsole as an application.

https://community.kde.org/Plasma/Notifications#DesktopEntry_in_notifyrc

Test Plan:
It seems the part stuff relies on the application (e.g. Yakuake) having their own notifyrc file, so associating this notifyrc with Konsole should be fine.

Could go into stable, it's just a field that is ignored by everybody else
{F6766670}

Reviewers: tcanabrava, hindenburg, ngraham

Reviewed By: ngraham

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D20470