Fix New tab behavior
Summary:
Fix the regression about the placement of a new tab:
it is now added either at the end, or next to the
current tab depending on the setting.

BUG: 406409

Test Plan:
In Konsole settings, select 'Put new tab after the
current tab' for the New tab behavior setting.
Open some new tabs and check they are opened at the
proper position.

Reviewers: #konsole, hindenburg, tcanabrava

Reviewed By: tcanabrava

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D20836