Return early if no Splitter
Historically Konsole is a complex software. there are many ways
to close a session, and one of the least used ones is to delete
the View where the session is, and deleting the session later.

Because the session will also try to delete the view when it's
being deleted, this led to a crash in yakuake.