Fix-ish scrollbar background with Qt Fusion style
Summary:
The "standard"/fallback Fusion style for Qt uses
QStyleHelper::backgroundColor() to determine if the
scrollbar is drawn on has a dark background.
QStyleHelper::backgroundColor() handles scrollbars
differently from everything else, and tries to get the
palette of the parent widget, and then gets the
QPalette::Base color of that (that probably makes sense
for someone other than me).

I think the scrollbar in the Qt Fusion style has way too
low contrast in the first place, but this way we should
at least invoke the right code in qfusionstyle, so it
can't blame konsole for it.

Test Plan:
Break with gdb at qfusionstyle.cpp:2539, verify that
isDarkBg is true.

Reviewers: #konsole, hindenburg, tcanabrava

Reviewed By: tcanabrava

Subscribers: ngraham, konsole-devel, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D19482