Relax rules for split selection via keyboard
Summary:
The idea is that a SplitterHeaderBar will be added in the
near future so the checks for the parent widget could be
wrong.
If instead of checking for a specific widget we check for
anything and walk to the parent list untill we find
a TerminalWidget seems sane, and safer.

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D21301