Variables for column/row are switched for Profile-Initial Terminal Size
Summary:
This dialogue page has two spin boxes setting the initial terminal size.
They are no longer individually labelled, but the first one has suffix
"rows" and the second has "columns".  However, they actually set
the columns and rows respectively.  The default settings are 80 for
the first box and 24 for the second, which are obviously the column
and row settings.

Test Plan:
Built Konsole with this change.  Checked that the spin box suffixes
are "columns" and 'rows" respectively and agree with what they set.

Reviewers: #konsole

Subscribers: konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D21516