Fix the logic for Maximize / Minimize
Summary:
The old logic is good but the splitters don't hide
themselves if all children is hidden, so we need to
hide them again.

Test Plan: The terminals seems a bit "cut" after restoration (this actually hits me with anything that shows / hides a session. @hindenburg  any hint to fix this behavior completely?

Reviewers: ngraham, hindenburg, #konsole

Subscribers: hindenburg, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D21308