Add a TerminalHeader when using splits
Summary:
The idea is that a SplitterHeaderBar will be added in the
near future so the checks for the parent widget could be
wrong.
If instead of checking for a specific widget we check for
anything and walk to the parent list untill we find
a TerminalWidget seems sane, and safer.

Phabricator messed with this review and send the wrong diff together (because of the same parent)
so I'm rephrasing it.

Reviewers: #konsole, hindenburg, ngraham, #vdg

Subscribers: aaronhoneycutt, kvermette, thsurrel, rizzitello, mart, hindenburg, gennad, fabianr, ndavis, shubham, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D21301