Allow loading of pre 19.08 session files
The format of the Konsole session file changed with 19.08. This was
overlooked and pre 19.08 files caused crashes.  That crash was fixed
in 19.08.1 to just start the default profile.
This patch will allow pre 19.08 session files to be loaded correctly.

BUG: 412173
FIXED-IN: 19.08.2
(cherry picked from commit 2de43394c98364edd48fb74bbaa467d5d281d445)