Fix tabs not indicating activity
Summary:
This fixes the bug with tabs not indicating terminal activity.
When the tab is split and has multiple child terminals, activity in any
of the child terminals will be indicated.

BUG: 406828
FIXED-IN: 19.12.0

Test Plan:
Open two tabs, run  `sleep 2` in the first one and switch to the other.
After the sleep ends, the tab activity is indicated just as it was before
the bug was introduced.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, anthonyfieroni, nucleo, konsole-devel, #konsole

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D25151