Paste urls of local files as a list of quoted strings
When pasting urls of local files, remove the the scheme part, "file://",
and paste the urls as a space-separated list of local file paths, which
are quoted as needed.

Commands such as cp, mv, ls... etc, don't understand/parse the file://
bit of local urls, therefore it's more useful to strip the scheme for local
files.

CCBUG: 408813