Copy _colorRandomization when copying a new ColorScheme
Summary:
The `_colorRandomization` property is new in 19.12, and should be
copied by the copy constructor as well.

This fixes the problem that it was not possible to disable the
"Randomly adjust colors for each session" option again once it was
enabled.

CCBUG: 415242
FIXED-IN: 19.12.1

Test Plan:
In the color scheme editor:
Choose "Black on Random Light" and click on Edit, disable
"Randomly adjust colors for each session", and click OK.
It's saved correctly now, previously it stayed on.

Alternatively, select some other color scheme, enable the option and
click OK. Then try to disable it again. This works now, previously it
also stayed on.

Select an existing colorscheme and click on New, the new color scheme
still properly inherits the "Randomly adjust colors for each session"
setting.

Reviewers: #konsole, hindenburg

Reviewed By: #konsole, hindenburg

Subscribers: hindenburg, konsole-devel

Tags: #konsole

Differential Revision: https://phabricator.kde.org/D26439

(cherry picked from commit 771c1733fd5ee9cfb854212b83df6de148608a65)