Fix context menu actions when search bar is enabled
The actions are destroyed during a focusIn / focusOut, and recreated
with the same content but different memory positions, and that
killed the QActions inside of the menu.

To trigger this is simple, open the search bar, type some url in
konsole, and right click on it: you will not see the actions
to copy and go to url.

moving the related code to QSharedPointer fixes this by delaying
the destruction of the pointer untill the menu is closed.