Add an option to set the custom color of text under the cursor
Current version of Konsole allows to select the custom color of the
cursor (Settings -> Edit Current Profile -> Appearance -> Cursor ->
Custom cursor color). This works well for I-Beam and Underline cursor
shapes. But for block-shaped cursors the text under the cursor is often
hard to read, e.g. bright text under the bright cursor or dark text
under the dark cursor.

This patch adds an ability to select the custom color of the character
under the cursor. The character uses this color only when the cursor
shape is Block and Konsole window is focused. Otherwise the character
uses its foreground color as usual.

FEATURE: 407915
FIXED-IN: 20.04