[FileFilter] Create HotSpot's for files in child dirs too
When creating HotSpot's for local files, create them for files in the
current dir and for files in sub-directories too.

Re-format the code used to build the regex pattern for more readability
(easier for seeing what the regex pattern will look like, I hope).

Use a static QRegularExpression object, so as to only construct it once,
the pattern doesn't change and this should help with performance. Also
make createFileRegex() not static.

Don't convert QList<QString> to QSet<QString>, without actual benchmarking
it could be that QList is actually efficient here.

Use const where appropriate.

BUG: 416376