[BookmarkMenu] Adapt the code to KBookmarks 5.69 changes
Starting from KBookmarks 5.69 the ctor that took a KActionCollection
parameter has been deprecated, and we need to manually add the various
bookmark actions to the actionCollection so that they show up in the
shortcusts editor.

Move the code that overrides the add bookmark shortcut, Ctrl+B, to the
BookmarkMenu class, so that all the relevant code is in one place.

BUG: 419981
(cherry picked from commit 6db7f8d2593ccb0238b2b547ed6eaf7da6a26c4c)