Add Split Settings: Toggle Visiblity and Dragger Size
The Splitter is a great addition to konsole, but it lacked one
thing that always made KDE software proud: Configurability.
Some people like to have the header bar, some people don't,
some people prefer to have thin lines separating the views, some people
prefer to have the lines a bit thicker so it's easy to drag them with
the mouse.

This patch introduces those two settings in a new Tab on the Settings.

FIXED-IN: 20.08

FEATURE:

GUI:

CHANGELOG: Add Splitter settings for visibility and drag size