[SessionController] Fix crash caused by text encoding menu
QMenu since 5.15 is hidden when an action is triggered, this caused a
crash in Konsole when trying to access the text encoding menu.

Now Session emits a signal when the text encoding is changed, the
SessionController can connect to that singal to set the current codec in
the KCodecAction object.

Also fix the EditProfileDialog so that when the KCodecAction menu is
shown the currently set codec is selected.

BUG: 419526

FIXED-IN: 20.08