Fix SearchTask memleak
There's a `autoDelete` option for a task, but it's up to the
task to implement how it will delete itself. For SearchHistoryTask
this never happened.

I tested holding the `search` button for around one minute and the
memory was slowly growing and never freed. Adding a destructor with a
debug to remove any doubt made me realize: The search task was only
removed when konsole closed.

This is not just about the Memory Management, as the search was also
connected to multiple objects and could potentially mishandle searches.

(cherry picked from commit 9372cf714753fee59c2c4748626635ddca4cc8e2)