Add thumbnails for certain file types on mouse hover
This allows for a user configured thumbnail image to be displayed
when the mouse is hovering over a file link.  Any file that KIO can
transform into a thumbnail (image, video, folder) will be displayed.

Simply move the mouse to a file while holding an user selected keypress
(Alt, Shift, Control or a combination of them).  The default
requires no key press.  The profile setting 'Underline files' much be
enabled for this to work.

https://invent.kde.org/utilities/konsole/-/merge_requests/93

FIXED-IN: 20.08
FEATURE:
GUI:
CHANGELOG: Add thumbnails for certain file types on mouse hover