Correctly handle selection and opening via clicks
Before we opened a link via mousePress but if we want to make a
selection that will open a file and drag.

Now we check if there is a selection before trying to open something.
So if you click and drag a url, it will not open. If you directly
click and release without dragging, it will open.

FIXED-IN: 20.08

https://invent.kde.org/utilities/konsole/-/merge_requests/9