Increase default terminal size
Note that the values I chose are a bit random. I just picked what I found
visually appealing (aspect ratio wise) and tried not to choose a to big
size, since I don't want to break things on smaller resolutions or annoy
people who do not like the terminal overlapping to much space.

With this values, Konsole uses about a quarter of the screen on a FHD
resolution, which I consider a standard resolution nowadays.

For more discussion about this also have a look at the relevant MR where
it has been reacted to some comments mentioning possible problems:
https://invent.kde.org/utilities/konsole/-/merge_requests/125