Add hint to also translate a leading space
Having spaces (especially leading or trailing ones) in translatable
strings is probably not a pretty good idea as that is a source of
errors. Maybe there can be some refactoring of the setSuffix function
to add a space by itself, so it can be left out of the translatable
string.

For now I added a hint that this space has to be translated as well.
I ran into this because it is not there in the German translation for
"row(s)".

https://invent.kde.org/utilities/konsole/-/merge_requests/126