HotSpot: move input events implementation to the base class
Then if one of the classes that inherit HotSpot need to reimplement an
input event, it can do so. I think this way we get less code duplication,
and if we need to change those methods, there is a chance only changing the
base implementation is necessary.

This also fixes file filter.

Initialise  member variables in FilterChain ctor, particularly initialising
_showUrlHint prevents an issue where url hints are displayed when moving
the mouse for the first time in a newly opened konsole window even if no
url hint modifiers are pressed.

In keyPressEvent it seems calling update() on the TerminalDisplay isn't
needed.