Revert "Correctly fix reverse url numering"
The previous behavior was actually the intended one, the URL hints are
intended to strictly increase from start to finish (physically on the
keyboard).

It might not be "technically" correct, but that's the original behavior
before the refactor (and I think it makes the most sense, that's why I
implemented it that way originally).

Also fixes activating the first URL, activating 0 didn't work.

This reverts commit 28e1a3e634236bf271cab8fa7ef4178656c9f506.

BUG: 427293